requirements.txt
openai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pillow>=10.0.0
pytest>=8.0.0
